chore: ui fixes added
