+++ web/components/SessionSidebar.workspace.test.mjs
@@ -2,21 +2,25 @@ import assert from "node:assert/strict";
 import { readFile } from "node:fs/promises";
 import test from "node:test";
 
-const source = await readFile(new URL("./SessionSidebar.tsx", import.meta.url), "utf8");
-const sessionItemSource = source.slice(source.indexOf("function SessionItem("));
+const url = (p) => new URL(`./session-sidebar/${p}`, import.meta.url);
+const modelSource = await readFile(url("use-session-sidebar-model.ts"), "utf8");
+const workspacePanelSource = await readFile(url("workspace-panel.tsx"), "utf8");
+const sessionItemSource = await readFile(url("session-item.tsx"), "utf8");
+const sessionItem = sessionItemSource.slice(sessionItemSource.indexOf("function SessionItem("));
+const sessionTreeItemSource = sessionItemSource.slice(sessionItemSource.indexOf("function SessionTreeItem("));
 
 function callbackBody(name, endMarker) {
-  const start = source.indexOf(`const ${name} = useCallback`);
-  const end = source.indexOf(endMarker, start);
+  const start = modelSource.indexOf(`const ${name} = useCallback`);
+  const end = modelSource.indexOf(endMarker, start);
   assert.notEqual(start, -1, `${name} callback not found`);
   assert.notEqual(end, -1, `${name} callback end not found`);
-  return source.slice(start, end);
+  return modelSource.slice(start, end);
 }
 
 test("sidebar no longer owns folder validation", () => {
-  assert.doesNotMatch(source, /\/api\/cwd\/validate/);
-  assert.doesNotMatch(source, /commitCustomPath/);
-  assert.doesNotMatch(source, /setValidatedProject\(/, "no sidebar-owned validatedProject state");
+  assert.doesNotMatch(modelSource, /\/api\/cwd\/validate/);
+  assert.doesNotMatch(modelSource, /commitCustomPath/);
+  assert.doesNotMatch(modelSource, /setValidatedProject\(/, "no sidebar-owned validatedProject state");
 });
 
 test("projectFor prefers the shell-provided validated identity before session and worktree fallbacks", () => {
@@ -30,65 +34,65 @@ test("projectFor prefers the shell-provided validated identity before session an
 });
 
 test("workspace browser uses native controls without partial tree or menu roles", () => {
-  assert.doesNotMatch(source, /role="tree"/);
-  assert.doesNotMatch(source, /role="treeitem"/);
-  assert.doesNotMatch(source, /role="menu"/);
-  assert.doesNotMatch(source, /role="menuitem"/);
-  assert.match(source, /<div className="workspace-row"/);
-  assert.match(source, /aria-expanded=\{isExpanded\} aria-controls=\{groupId\}/);
-  assert.match(source, /aria-current=\{isSelected \? "page" : undefined\}/);
-  assert.match(source, /aria-label=\{t\("sidebar\.workspaces"\)\}/);
+  assert.doesNotMatch(workspacePanelSource, /role="tree"/);
+  assert.doesNotMatch(workspacePanelSource, /role="treeitem"/);
+  assert.doesNotMatch(workspacePanelSource, /role="menu"/);
+  assert.doesNotMatch(workspacePanelSource, /role="menuitem"/);
+  assert.match(workspacePanelSource, /<div className="workspace-row"/);
+  assert.match(workspacePanelSource, /aria-expanded=\{isExpanded\} aria-controls=\{groupId\}/);
+  assert.match(workspacePanelSource, /aria-current=\{isSelected \? "page" : undefined\}/);
+  assert.match(workspacePanelSource, /aria-label=\{t\("sidebar\.workspaces"\)\}/);
 });
 
 test("session select is a native button without nested interactive controls", () => {
-  assert.match(sessionItemSource, /<button type="button" className="session-row-main"/);
-  const mainStart = sessionItemSource.indexOf('<button type="button" className="session-row-main"');
-  const tagEnd = sessionItemSource.indexOf(">", mainStart);
-  const mainEnd = sessionItemSource.indexOf("</button>", mainStart);
+  assert.match(sessionItem, /<button type="button" className="session-row-main"/);
+  const mainStart = sessionItem.indexOf('<button type="button" className="session-row-main"');
+  const tagEnd = sessionItem.indexOf(">", mainStart);
+  const mainEnd = sessionItem.indexOf("</button>", mainStart);
   assert.ok(mainStart >= 0 && mainEnd > mainStart, "main session select button found");
-  assert.doesNotMatch(sessionItemSource.slice(tagEnd, mainEnd), /<button/, "no nested button inside the session select button");
+  assert.doesNotMatch(sessionItem.slice(tagEnd, mainEnd), /<button/, "no nested button inside the session select button");
 });
 
 test("search drives searchWorkspaces with selected, running, and unread context", () => {
   assert.match(
-    source,
-    /searchWorkspaces\(workspaceInputs, workspaceQuery, \{[\s\S]*?selectedWorkspaceKey: selectedProject\?\.key,[\s\S]*?selectedSessionId,[\s\S]*?runningSessionIds,[\s\S]*?unreadSessionIds,\s*\}\)/,
+    modelSource,
+    /searchWorkspaces\(workspaceInputs, workspaceQuery, \{[\s\S]*?selectedWorkspaceKey: projectFor\(explorerCwd\)\?\.key,[\s\S]*?selectedSessionId,[\s\S]*?runningSessionIds,[\s\S]*?unreadSessionIds,\s*\}\)/,
   );
 });
 
 test("query expansion is effective and never writes the stored expansion set", () => {
-  assert.match(source, /row\.hasQueryMatch \|\| expandedWorkspaceKeys\.has\(row\.key\)/);
-  const searchBlock = source.slice(
-    source.indexOf("const visibleWorkspaces"),
-    source.indexOf("const hasQueryMatches"),
+  assert.match(workspacePanelSource, /row\.hasQueryMatch \|\| expandedWorkspaceKeys\.has\(row\.key\)/);
+  const searchBlock = modelSource.slice(
+    modelSource.indexOf("const visibleWorkspaces"),
+    modelSource.indexOf("const hasQueryMatches"),
   );
   assert.doesNotMatch(searchBlock, /setExpandedWorkspaceKeys/);
 });
 
 test("search force-expands fork hierarchies through the recursive rows", () => {
-  assert.match(source, /forceExpanded=\{/);
-  assert.match(source, /\{hasChildren && \(forceExpanded \|\| !collapsed\) && \(/);
+  assert.match(sessionTreeItemSource, /forceExpanded=\{/);
+  assert.match(sessionTreeItemSource, /\{hasChildren && \(forceExpanded \|\| !collapsed\) && \(/);
 });
 
 test("empty search result offers clearing the query", () => {
-  assert.match(
-    source,
-    /<div role="status"[\s\S]*?t\("sidebar\.noWorkspaceMatches"\)[\s\S]*?onClick=\{\(\) => setWorkspaceQuery\(""\)\}[\s\S]*?t\("sidebar\.clearSearch"\)/,
-  );
+  assert.match(workspacePanelSource, /<div role="status"/);
+  assert.match(workspacePanelSource, /onClick=\{\(\) => model\.setWorkspaceQuery\(""\)\}/);
+  assert.match(workspacePanelSource, /clearLabel=\{t\("sidebar\.clearSearch"\)}/);
+  assert.match(workspacePanelSource, /message=\{t\("sidebar\.noWorkspaceMatches"\)}/);
 });
 
 test("workspace activity counts derive from the unfiltered session set", () => {
-  assert.match(source, /getProjectActivity\(allSessions, runningSessionIds, unreadSessionIds\)/);
+  assert.match(modelSource, /getProjectActivity\(allSessions, runningSessionIds, unreadSessionIds\)/);
 });
 
 test("the transient workspace row keeps the exact validated cwd", () => {
-  assert.match(source, /transientWorkspace\(validatedProject \?\? null, recentProjects\)/);
-  assert.match(source, /\.\.\.\(pending \? \[\{ \.\.\.pending, sessions: \[\] \} \] : \[\]\)/);
+  assert.match(modelSource, /transientWorkspace\(validatedProject \?\? null, recentProjects\)/);
+  assert.match(modelSource, /\.\.\.\(pending \? \[\{ \.\.\.pending, sessions: \[\] \}\] : \[\]\)/);
 });
 
 test("re-selecting the active workspace only ensures expansion and never rewrites cwd", () => {
-  const body = callbackBody("handleWorkspaceSelect", "\n  }, [");
-  const guardAt = body.indexOf("row.key === selectedProject?.key");
+  const body = callbackBody("handleWorkspaceSelect", "\n  },");
+  const guardAt = body.indexOf("row.key === explorerProject?.key");
   assert.ok(guardAt >= 0, "already-selected guard present");
   const between = body.slice(guardAt, body.indexOf("setSelectedCwd("));
   assert.doesNotMatch(between, /setSelectedCwd\(/);
@@ -96,18 +100,18 @@ test("re-selecting the active workspace only ensures expansion and never rewrite
 });
 
 test("worktree switcher renders inside the workspace list and collapse closes its dropdown", () => {
-  const listAt = source.indexOf("visibleWorkspaces.map(");
-  const switcherAt = source.indexOf("{showWorktreeSwitcher && (() => {");
-  assert.ok(listAt >= 0 && switcherAt > listAt, "worktree switcher moved inside the workspace rows");
+  const switcherAt = workspacePanelSource.indexOf("{showWorktreeSwitcher &&");
+  const sessionsAt = workspacePanelSource.indexOf("\"workspace-sessions\"");
+  assert.ok(sessionsAt >= 0 && switcherAt > sessionsAt, "worktree switcher moved inside the workspace sessions");
   assert.match(
-    source,
+    modelSource,
     /if \(selectedWorkspaceKey && !selectedRowVisible\) \{\s*setWtDropdownOpen\(false\);/,
   );
 });
 
 test("session action guards remain intact", () => {
-  assert.match(sessionItemSource, /if \(session\.transient\) return;/);
-  assert.match(sessionItemSource, /if \(e\.shiftKey\) \{/);
-  assert.match(sessionItemSource, /dispatchSessionRowContextMenu\(\{/);
-  assert.match(sessionItemSource, /onContextMenu=\{confirmDelete \|\| renaming \? undefined : handleContextMenu\}/);
+  assert.match(sessionItem, /if \(session\.transient\) return;/);
+  assert.match(sessionItem, /if \(e\.shiftKey\) \{/);
+  assert.match(sessionItem, /dispatchSessionRowContextMenu\(\{/);
+  assert.match(sessionItem, /onContextMenu=\{confirmDelete \|\| renaming \? undefined : handleContextMenu\}/);
 });
