+++ web/components/SessionSidebar.worktree.test.mjs
@@ -2,14 +2,15 @@ import assert from "node:assert/strict";
 import { readFile } from "node:fs/promises";
 import test from "node:test";
 
-const source = await readFile(new URL("./SessionSidebar.tsx", import.meta.url), "utf8");
+const url = (p) => new URL(`./session-sidebar/${p}`, import.meta.url);
+const modelSource = await readFile(url("use-session-sidebar-model.ts"), "utf8");
 
 test("uses the server-resolved current worktree identity", () => {
-  assert.match(source, /currentWorktreePath: string \| null/);
+  assert.match(modelSource, /currentWorktreePath: string \| null/);
   assert.match(
-    source,
+    modelSource,
     /const currentWorktree =[\s\S]*?worktreeState\.currentWorktreePath[\s\S]*?worktree\.path === worktreeState\.currentWorktreePath/,
   );
-  assert.match(source, /if \(currentWorktreePath === path\) setSelectedCwd\(worktreeState\.projectRoot\)/);
-  assert.doesNotMatch(source, /const isCurrent = wt\.path === selectedCwd/);
+  assert.match(modelSource, /if \(currentWorktreePath === path\) setSelectedCwd\(worktreeState\.projectRoot\)/);
+  assert.doesNotMatch(modelSource, /const isCurrent = wt\.path === selectedCwd/);
 });
