+++ web/components/session-sidebar/utils.tsx
@@ -1,0 +1,260 @@
+"use client";
+
+import { useEffect, useRef, useState, type CSSProperties, type ReactNode } from "react";
+import type { SessionInfo } from "@/lib/types";
+
+/** A session in the fork hierarchy shown in the workspace row. Distinct from
+ *  the message-entry `SessionTreeNode` in lib/types (BranchNavigator's tree). */
+export interface SessionItemNode {
+  session: SessionInfo;
+  children: SessionItemNode[];
+}
+
+declare global {
+  interface Window {
+    piDesktop?: {
+      selectDirectory: () => Promise<string | null>;
+    };
+  }
+}
+
+/* ─────────────────────────── Pure helpers ─────────────────────────── */
+
+const UNREAD_SESSIONS_STORAGE_KEY = "pi-web:unread-session-ids";
+
+export function loadUnreadSessionIds(): Set<string> {
+  if (typeof window === "undefined") return new Set();
+  try {
+    const raw = window.localStorage.getItem(UNREAD_SESSIONS_STORAGE_KEY);
+    if (!raw) return new Set();
+    const parsed = JSON.parse(raw) as unknown;
+    if (Array.isArray(parsed)) return new Set(parsed.filter((id): id is string => typeof id === "string"));
+    return new Set();
+  } catch {
+    return new Set();
+  }
+}
+
+export function saveUnreadSessionIds(ids: Set<string>): void {
+  if (typeof window === "undefined") return;
+  try {
+    if (ids.size === 0) window.localStorage.removeItem(UNREAD_SESSIONS_STORAGE_KEY);
+    else window.localStorage.setItem(UNREAD_SESSIONS_STORAGE_KEY, JSON.stringify([...ids]));
+  } catch {
+    // ignore storage quota / privacy-mode errors
+  }
+}
+
+export function formatRelativeTime(dateStr: string): string {
+  const date = new Date(dateStr);
+  const now = new Date();
+  const diff = now.getTime() - date.getTime();
+  const mins = Math.floor(diff / 60000);
+  const hours = Math.floor(diff / 3600000);
+  const days = Math.floor(diff / 86400000);
+  if (mins < 1) return "just now";
+  if (mins < 60) return `${mins}min`;
+  if (hours < 24) return `${hours}h`;
+  if (days < 7) return `${days}d`;
+  return date.toLocaleDateString();
+}
+
+/**
+ * Derive a clean, human label for a session title. When the raw title is an
+ * absolute filesystem path — the common case when a user pastes a file/dir as
+ * the first message — collapse it to its final path segment so the list reads
+ * as `Screenshot.png` instead of `/var/folders/3r/.../Screenshot.png`. Free
+ * text passes through unchanged.
+ */
+export function friendlySessionTitle(raw: string): string {
+  const trimmed = raw.trim();
+  const looksLikePath = (/^\//.test(trimmed) || /^[A-Za-z]:[\\/]/.test(trimmed)) && /[\\/]/.test(trimmed);
+  if (!looksLikePath) return trimmed;
+  const base = trimmed.split(/[\\/]/).filter(Boolean).pop() ?? trimmed;
+  return base;
+}
+
+/** Substitute the home dir prefix with ~ (no path truncation — see PathLabel) */
+export function displayCwd(cwd: string, homeDir?: string): string {
+  return (homeDir && cwd.startsWith(homeDir)) ? "~" + cwd.slice(homeDir.length) : cwd;
+}
+
+/**
+ * Path label that ellipsizes on the LEFT, keeping the (most relevant) trailing
+ * segments visible. The rtl container moves the ellipsis to the left edge; the
+ * inner plaintext bidi isolation keeps the path rendered strictly left-to-right.
+ */
+export function PathLabel({ text, style }: { text: string; style?: CSSProperties }) {
+  return (
+    <span
+      style={{
+        overflow: "hidden",
+        textOverflow: "ellipsis",
+        whiteSpace: "nowrap",
+        display: "block",
+        minWidth: 0,
+        lineHeight: 1.35,
+        direction: "rtl",
+        textAlign: "left",
+        ...style,
+      }}
+    >
+      <span style={{ unicodeBidi: "plaintext" }}>{text}</span>
+    </span>
+  );
+}
+
+export interface WorktreeEntry {
+  path: string;
+  branch: string | null;
+  isMain: boolean;
+}
+
+/** Build a sorted session tree (fork hierarchies) from a flat list. */
+export function buildSessionTree(sessions: SessionInfo[]): SessionItemNode[] {
+  const byId = new Map<string, SessionItemNode>();
+  for (const s of sessions) {
+    byId.set(s.id, { session: s, children: [] });
+  }
+
+  const parentOf = new Map<string, string>();
+  for (const s of sessions) {
+    if (s.parentSessionId) parentOf.set(s.id, s.parentSessionId);
+  }
+
+  function resolveAncestor(id: string): string | null {
+    let cur = parentOf.get(id);
+    const visited = new Set<string>();
+    while (cur) {
+      if (visited.has(cur)) return null; // cycle guard
+      visited.add(cur);
+      if (byId.has(cur)) return cur;
+      cur = parentOf.get(cur);
+    }
+    return null;
+  }
+
+  const roots: SessionItemNode[] = [];
+  for (const node of byId.values()) {
+    const ancestor = resolveAncestor(node.session.id);
+    if (ancestor) {
+      byId.get(ancestor)!.children.push(node);
+    } else {
+      roots.push(node);
+    }
+  }
+
+  const sort = (nodes: SessionItemNode[]) => {
+    nodes.sort((a, b) => b.session.modified.localeCompare(a.session.modified));
+    nodes.forEach((n) => sort(n.children));
+  };
+  sort(roots);
+  return roots;
+}
+
+/* ─────────────────────────── Small presentational bits ─────────────────────────── */
+
+const DROPDOWN_ANIMATION_MS = 140;
+
+export function AnimatedDropdown({ open, children, style }: { open: boolean; children: ReactNode; style: CSSProperties }) {
+  const [mounted, setMounted] = useState(open);
+  const [visible, setVisible] = useState(open);
+
+  useEffect(() => {
+    let frame: number | undefined;
+    let timeout: ReturnType<typeof setTimeout> | undefined;
+
+    if (open) {
+      setMounted(true);
+      setVisible(false);
+      frame = window.requestAnimationFrame(() => {
+        frame = window.requestAnimationFrame(() => setVisible(true));
+      });
+    } else {
+      setVisible(false);
+      timeout = setTimeout(() => setMounted(false), DROPDOWN_ANIMATION_MS);
+    }
+
+    return () => {
+      if (frame !== undefined) window.cancelAnimationFrame(frame);
+      if (timeout) clearTimeout(timeout);
+    };
+  }, [open]);
+
+  if (!mounted) return null;
+
+  return (
+    <div
+      style={{
+        ...style,
+        opacity: visible ? 1 : 0,
+        transform: visible ? "translateY(0) scale(1)" : "translateY(-8px) scale(0.96)",
+        transformOrigin: "top center",
+        transition: `opacity ${DROPDOWN_ANIMATION_MS}ms ease, transform ${DROPDOWN_ANIMATION_MS}ms ease`,
+        pointerEvents: open ? "auto" : "none",
+      }}
+    >
+      {children}
+    </div>
+  );
+}
+
+export function ToolbarIconButton({
+  onClick,
+  title,
+  disabled,
+  skipHover,
+  color,
+  background = "none",
+  marginRight,
+  ariaPressed,
+  children,
+}: {
+  onClick: () => void;
+  title: string;
+  disabled?: boolean;
+  skipHover?: boolean;
+  color: string;
+  background?: string;
+  marginRight?: number;
+  ariaPressed?: boolean;
+  children: ReactNode;
+}) {
+  const enter = (e: React.MouseEvent<HTMLButtonElement>) => {
+    if (disabled || skipHover) return;
+    e.currentTarget.style.color = "var(--text-muted)";
+    e.currentTarget.style.background = "var(--bg-hover)";
+  };
+  const leave = (e: React.MouseEvent<HTMLButtonElement>) => {
+    if (disabled || skipHover) return;
+    e.currentTarget.style.color = color;
+    e.currentTarget.style.background = background;
+  };
+  return (
+    <button
+      onClick={onClick}
+      disabled={disabled}
+      title={title}
+      aria-label={title}
+      aria-pressed={ariaPressed}
+      style={{
+        position: "relative",
+        display: "flex", alignItems: "center", justifyContent: "center",
+        width: 26, height: 26, padding: 0, marginRight,
+        background,
+        border: "none",
+        color,
+        cursor: disabled ? "default" : "pointer",
+        borderRadius: 5,
+        flexShrink: 0,
+        opacity: disabled ? 0.6 : 1,
+        transition: "color 0.3s, background 0.3s",
+      }}
+      onMouseEnter={enter}
+      onMouseLeave={leave}
+    >
+      {children}
+    </button>
+  );
+}
+
