+++ web/lib/i18n/messages/en.ts
@@ -89,6 +89,9 @@ export const enLocale: LocalePlugin = {
     "sidebar.new": "New",
     "sidebar.recents": "Recents",
     "sidebar.newSessionTitle": "New session in {path}",
+    "sidebar.collapse": "Collapse sidebar",
+    "sidebar.expand": "Expand sidebar",
+    "sidebar.newSession": "New session",
     "sidebar.workspaces": "Workspaces",
     "sidebar.searchWorkspaces": "Search workspaces and sessions…",
     "sidebar.clearSearch": "Clear search",
