+++ web/lib/i18n/messages/zh-CN.ts
@@ -89,6 +89,9 @@ export const zhCNLocale: LocalePlugin = {
     "sidebar.new": "新建",
     "sidebar.recents": "最近",
     "sidebar.newSessionTitle": "在 {path} 中新建会话",
+    "sidebar.collapse": "折叠侧边栏",
+    "sidebar.expand": "展开侧边栏",
+    "sidebar.newSession": "新建会话",
     "sidebar.workspaces": "工作区",
     "sidebar.searchWorkspaces": "搜索工作区和会话…",
     "sidebar.clearSearch": "清除搜索",
